refactor(difficulty modifier): extract modifier abstract class
