+++ src/main/java/world/anhgelus/architectsland/difficultydeathscaler/difficulty/modifier/BlockBreakSpeedModifier.java
@@ -6,9 +6,8 @@ import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
-import world.anhgelus.architectsland.difficultydeathscaler.difficulty.DifficultyManager;
 
-public class BlockBreakSpeedModifier extends DifficultyManager.Modifier<ServerPlayerEntity> {
+public class BlockBreakSpeedModifier extends Modifier<ServerPlayerEntity> {
     public static final RegistryEntry<EntityAttribute> ATTRIBUTE = EntityAttributes.PLAYER_BLOCK_BREAK_SPEED;
     public static final EntityAttributeModifier.Operation OPERATION = EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE;
     protected static Identifier ID = Identifier.of(PREFIX + "block_break_speed_modifier");
