+++ src/main/java/world/anhgelus/architectsland/difficultydeathscaler/difficulty/modifier/FollowRangeModifier.java
@@ -4,12 +4,11 @@ import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.mob.HostileEntity;
+import net.minecraft.entity.mob.ZombieEntity;
 import net.minecraft.registry.entry.RegistryEntry;
-import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
-import world.anhgelus.architectsland.difficultydeathscaler.difficulty.DifficultyManager;
 
-public class FollowRangeModifier extends DifficultyManager.Modifier<HostileEntity> {
+public class FollowRangeModifier extends Modifier<HostileEntity> {
     public static final RegistryEntry<EntityAttribute> ATTRIBUTE = EntityAttributes.GENERIC_FOLLOW_RANGE;
     public static final EntityAttributeModifier.Operation OPERATION = EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE;
     protected static Identifier ID = Identifier.of(PREFIX + "follow_range_modifier");
