+++ src/main/java/world/anhgelus/architectsland/difficultydeathscaler/difficulty/modifier/PlayerHealthModifier.java
@@ -6,9 +6,8 @@ import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
-import world.anhgelus.architectsland.difficultydeathscaler.difficulty.DifficultyManager;
 
-public abstract class PlayerHealthModifier extends DifficultyManager.Modifier<ServerPlayerEntity> {
+public abstract class PlayerHealthModifier extends Modifier<ServerPlayerEntity> {
     public static final RegistryEntry<EntityAttribute> ATTRIBUTE = EntityAttributes.GENERIC_MAX_HEALTH;
     public static final EntityAttributeModifier.Operation OPERATION = EntityAttributeModifier.Operation.ADD_VALUE;
