+++ src/main/java/world/anhgelus/architectsland/difficultydeathscaler/difficulty/modifier/StepHeightModifier.java
@@ -5,11 +5,9 @@ import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.mob.HostileEntity;
 import net.minecraft.registry.entry.RegistryEntry;
-import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
-import world.anhgelus.architectsland.difficultydeathscaler.difficulty.DifficultyManager;
 
-public class StepHeightModifier extends DifficultyManager.Modifier<HostileEntity> {
+public class StepHeightModifier extends Modifier<HostileEntity> {
     public static final RegistryEntry<EntityAttribute> ATTRIBUTE = EntityAttributes.GENERIC_STEP_HEIGHT;
     public static final EntityAttributeModifier.Operation OPERATION = EntityAttributeModifier.Operation.ADD_VALUE;
     protected static Identifier ID = Identifier.of(PREFIX + "step_height_modifier");
